+++ packages/discovery-react-components/src/components/DocumentPreview/components/PdfHighlight/utils/textLayout/index.ts
@@ -1,0 +1,3 @@
+export { HtmlBboxTextLayout } from './HtmlBboxTextLayout';
+export { PdfTextContentTextLayout } from './PdfTextContentTextLayout';
+export { TextMappingsTextLayout } from './TextMappingsTextLayout';